SwiftExtension // Add String.charDescriptions: [String], etc. (#249)
* SwiftExtension // Add String.charDescriptions: [String].

* SwiftExtension // Including certain SwiftUI extensions.
