+++ Source/Modules/UIModules/PrefUI/CtlPrefUI.swift
@@ -7,6 +7,7 @@
 // requirements defined in MIT License.
 
 import SSPreferences
+import SwiftExtension
 import SwiftUI
 
 extension NSImage {
@@ -126,73 +127,4 @@ class CtlPrefUI {
     style: .toolbarItems
   )
   static let shared = CtlPrefUI()
-}
-
-// MARK: - Add "onChange" support.
-
-// Ref: https://mjeld.com/swiftui-macos-10-15-toggle-onchange/
-
-@available(macOS 10.15, *)
-extension Binding {
-  public func onChange(_ action: @escaping () -> Void) -> Binding {
-    Binding(
-      get: {
-        wrappedValue
-      },
-      set: { newValue in
-        wrappedValue = newValue
-        action()
-      }
-    )
-  }
-}
-
-// MARK: - Add ".tooltip" support.
-
-// Ref: https://stackoverflow.com/a/63217861
-
-@available(macOS 10.15, *)
-@available(macOS, obsoleted: 11)
-struct Tooltip: NSViewRepresentable {
-  let tooltip: String
-
-  func makeNSView(context _: NSViewRepresentableContext<Tooltip>) -> NSView {
-    let view = NSView()
-    view.toolTip = tooltip
-
-    return view
-  }
-
-  func updateNSView(_: NSView, context _: NSViewRepresentableContext<Tooltip>) {}
-}
-
-@available(macOS 10.15, *)
-@available(macOS, obsoleted: 11)
-extension View {
-  public func help(_ tooltip: String) -> some View {
-    overlay(Tooltip(tooltip: tooltip))
-  }
-}
-
-// MARK: - Windows Aero in Swift UI
-
-// Ref: https://stackoverflow.com/questions/62461957
-
-@available(macOS 10.15, *)
-struct VisualEffectView: NSViewRepresentable {
-  let material: NSVisualEffectView.Material
-  let blendingMode: NSVisualEffectView.BlendingMode
-
-  func makeNSView(context _: Context) -> NSVisualEffectView {
-    let visualEffectView = NSVisualEffectView()
-    visualEffectView.material = material
-    visualEffectView.blendingMode = blendingMode
-    visualEffectView.state = NSVisualEffectView.State.active
-    return visualEffectView
-  }
-
-  func updateNSView(_ visualEffectView: NSVisualEffectView, context _: Context) {
-    visualEffectView.material = material
-    visualEffectView.blendingMode = blendingMode
-  }
 }
