+++ Source/Modules/UIModules/PrefUI/VwrPrefPaneDictionary.swift
@@ -9,6 +9,7 @@
 import BookmarkManager
 import SSPreferences
 import Shared
+import SwiftExtension
 import SwiftUI
 
 @available(macOS 10.15, *)
