+++ Source/Modules/UIModules/PrefUI/VwrPrefPaneKeyboard.swift
@@ -9,6 +9,7 @@
 import IMKUtils
 import SSPreferences
 import Shared
+import SwiftExtension
 import SwiftUI
 
 @available(macOS 10.15, *)
